+++ Sources/Layout/MessagesCollectionViewFlowLayout.swift
@@ -118,10 +118,8 @@ open class MessagesCollectionViewFlowLayout: UICollectionViewFlowLayout {
         guard let attributesArray = super.layoutAttributesForElements(in: rect) as? [MessagesCollectionViewLayoutAttributes] else {
             return nil
         }
-        for attributes in attributesArray {
-            if attributes.representedElementCategory == .cell {
-                configure(attributes: attributes)
-            }
+        for attributes in attributesArray where attributes.representedElementCategory == .cell {
+            configure(attributes: attributes)
         }
         return attributesArray
     }
@@ -129,289 +127,12 @@ open class MessagesCollectionViewFlowLayout: UICollectionViewFlowLayout {
     open override func layoutAttributesForItem(at indexPath: IndexPath) -> UICollectionViewLayoutAttributes? {
         guard let attributes = super.layoutAttributesForItem(at: indexPath) as? MessagesCollectionViewLayoutAttributes else {
             return nil
-
         }
         if attributes.representedElementCategory == .cell {
             configure(attributes: attributes)
         }
         return attributes
     }
-
-    // MARK: - Layout Invalidation
-
-    open override func shouldInvalidateLayout(forBoundsChange newBounds: CGRect) -> Bool {
-        if collectionView?.bounds.width != newBounds.width {
-            removeAllCachedAttributes()
-            return true
-        } else {
-            return false
-        }
-    }
-
-    open override func invalidationContext(forBoundsChange newBounds: CGRect) -> UICollectionViewLayoutInvalidationContext {
-        let context = super.invalidationContext(forBoundsChange: newBounds)
-        guard let flowLayoutContext = context as? UICollectionViewFlowLayoutInvalidationContext else { return context }
-        flowLayoutContext.invalidateFlowLayoutDelegateMetrics = shouldInvalidateLayout(forBoundsChange: newBounds)
-        return flowLayoutContext
-    }
-
-    /// Removes the cached layout information for all `MessageType`s.
-    public func removeAllCachedAttributes() {
-        textMessageAttributesCache.removeAllObjects()
-        attributedTextMessageAttributesCache.removeAllObjects()
-        emojiMessageAttributesCache.removeAllObjects()
-        photoMessageAttributesCache.removeAllObjects()
-        videoMessageAttributesCache.removeAllObjects()
-        locationMessageAttributesCache.removeAllObjects()
-        customMessageAttributesCache.removeAllObjects()
-    }
-
-    public func removeAllCachedAttributes(for messageType: MessageData) {
-        let cache = attributeCache(for: messageType)
-        cache.removeAllObjects()
-    }
-
-    public func removeCachedAttributes(for message: MessageType) {
-        let cache = attributeCache(for: message.data)
-        cache.removeObject(forKey: message.messageId as NSString)
-    }
-
-    public func removeCachedAttributes(for messageId: String) {
-        textMessageAttributesCache.removeObject(forKey: messageId as NSString)
-        attributedTextMessageAttributesCache.removeObject(forKey: messageId as NSString)
-        emojiMessageAttributesCache.removeObject(forKey: messageId as NSString)
-        photoMessageAttributesCache.removeObject(forKey: messageId as NSString)
-        videoMessageAttributesCache.removeObject(forKey: messageId as NSString)
-        locationMessageAttributesCache.removeObject(forKey: messageId as NSString)
-        customMessageAttributesCache.removeObject(forKey: messageId as NSString)
-    }
-
-    @objc
-    private func handleOrientationChange(_ notification: Notification) {
-        removeAllCachedAttributes()
-        invalidateLayout()
-    }
-
-    // MARK: - Avatar Size
-
-    internal func avatarPosition(for message: MessageType, at indexPath: IndexPath) -> AvatarPosition {
-        let layout = layoutForMessage(message: message)
-        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
-        var position = isFromCurrentSender ? layout.outgoingAvatarPosition : layout.incomingAvatarPosition
-
-        switch position.horizontal {
-        case .cellTrailing, .cellLeading:
-            break
-        case .natural:
-            position.horizontal = isFromCurrentSender ? .cellTrailing : .cellLeading
-        }
-        return position
-    }
-
-    internal func avatarSize(for message: MessageType, at indexPath: IndexPath) -> CGSize {
-        let layout = layoutForMessage(message: message)
-        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
-        return isFromCurrentSender ? layout.outgoingAvatarSize : layout.incomingAvatarSize
-    }
-
-    // MARK: - Cell Top Label Size
-
-    internal func cellTopLabelAlignment(for message: MessageType, at indexPath: IndexPath) -> LabelAlignment {
-        let layout = layoutForMessage(message: message)
-        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
-        return isFromCurrentSender ? layout.outgoingCellTopLabelAlignment : layout.incomingCellTopLabelAlignment
-    }
-
-    internal func cellTopLabelSize(for message: MessageType, at indexPath: IndexPath) -> CGSize {
-        let text = messagesDataSource.cellTopLabelAttributedText(for: message, at: indexPath)
-        guard let topLabelText = text else { return .zero }
-        let maxWidth = cellTopLabelMaxWidth(for: message, at: indexPath)
-        return labelSize(for: topLabelText, considering: maxWidth)
-    }
-
-    internal func cellTopLabelMaxWidth(for message: MessageType, at indexPath: IndexPath) -> CGFloat {
-        let alignment = cellTopLabelAlignment(for: message, at: indexPath)
-        let position = avatarPosition(for: message, at: indexPath)
-        let avatarWidth = avatarSize(for: message, at: indexPath).width
-        let containerSize = messageContainerSize(for: message, at: indexPath)
-        let messagePadding = messageContainerPadding(for: message, at: indexPath)
-
-        let avatarHorizontal = position.horizontal
-        let avatarVertical = position.vertical
-
-        switch (alignment, avatarHorizontal) {
-
-        case (.cellLeading, _), (.cellTrailing, _):
-            let width = itemWidth - alignment.insets.horizontal
-            return avatarVertical != .cellTop ? width : width - avatarWidth
-
-        case (.cellCenter, _):
-            let width = itemWidth - alignment.insets.horizontal
-            return avatarVertical != .cellTop ? width : width - (avatarWidth * 2)
-
-        case (.messageTrailing, .cellLeading):
-            let width = containerSize.width + messagePadding.left - alignment.insets.horizontal
-            return avatarVertical == .cellTop ? width : width + avatarWidth
-
-        case (.messageLeading, .cellTrailing):
-            let width = containerSize.width + messagePadding.right - alignment.insets.horizontal
-            return avatarVertical == .cellTop ? width : width + avatarWidth
-
-        case (.messageLeading, .cellLeading):
-            return itemWidth - avatarWidth - messagePadding.left - alignment.insets.horizontal
-
-        case (.messageTrailing, .cellTrailing):
-            return itemWidth - avatarWidth - messagePadding.right - alignment.insets.horizontal
-
-        case (_, .natural):
-            fatalError(MessageKitError.avatarPositionUnresolved)
-        }
-    }
-
-    // MARK: - Cell Bottom Label Size
-
-    internal func cellBottomLabelAlignment(for message: MessageType, at indexPath: IndexPath) -> LabelAlignment {
-        let layout = layoutForMessage(message: message)
-        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
-        return isFromCurrentSender ? layout.outgoingCellBottomLabelAlignment : layout.incomingCellBottomLabelAlignment
-    }
-
-    internal func cellBottomLabelSize(for message: MessageType, at indexPath: IndexPath) -> CGSize {
-        let text = messagesDataSource.cellBottomLabelAttributedText(for: message, at: indexPath)
-        guard let bottomLabelText = text else { return .zero }
-        let maxWidth = cellBottomLabelMaxWidth(for: message, at: indexPath)
-        return labelSize(for: bottomLabelText, considering: maxWidth)
-    }
-
-    internal func cellBottomLabelMaxWidth(for message: MessageType, at indexPath: IndexPath) -> CGFloat {
-
-        let alignment = cellBottomLabelAlignment(for: message, at: indexPath)
-        let avatarWidth = avatarSize(for: message, at: indexPath).width
-        let containerSize = messageContainerSize(for: message, at: indexPath)
-        let messagePadding = messageContainerPadding(for: message, at: indexPath)
-        let position = avatarPosition(for: message, at: indexPath)
-
-        let avatarHorizontal = position.horizontal
-        let avatarVertical = position.vertical
-
-        switch (alignment, avatarHorizontal) {
-
-        case (.cellLeading, _), (.cellTrailing, _):
-            let width = itemWidth - alignment.insets.horizontal
-            return avatarVertical != .cellBottom ? width : width - avatarWidth
-
-        case (.cellCenter, _):
-            let width = itemWidth - alignment.insets.horizontal
-            return avatarVertical != .cellBottom ? width : width - (avatarWidth * 2)
-
-        case (.messageTrailing, .cellLeading):
-            let width = containerSize.width + messagePadding.left - alignment.insets.horizontal
-            return avatarVertical == .cellBottom ? width : width + avatarWidth
-
-        case (.messageLeading, .cellTrailing):
-            let width = containerSize.width + messagePadding.right - alignment.insets.horizontal
-            return avatarVertical == .cellBottom ? width : width + avatarWidth
-
-        case (.messageLeading, .cellLeading):
-            return itemWidth - avatarWidth - messagePadding.left - alignment.insets.horizontal
-
-        case (.messageTrailing, .cellTrailing):
-            return itemWidth - avatarWidth - messagePadding.right - alignment.insets.horizontal
-
-        case (_, .natural):
-            fatalError(MessageKitError.avatarPositionUnresolved)
-        }
-    }
-
-    // MARK: - Message Container Size
-
-    internal func messageLabelInsets(for message: MessageType, at indexPath: IndexPath) -> UIEdgeInsets {
-        guard let layout = layoutForMessage(message: message) as? TextMessageLayout else {
-            fatalError("Layout is not TextMessageLayout")
-        }
-        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
-        return isFromCurrentSender ? layout.outgoingMessageLabelInsets : layout.incomingMessageLabelInsets
-    }
-
-    internal func messageContainerPadding(for message: MessageType, at indexPath: IndexPath) -> UIEdgeInsets {
-        let layout = layoutForMessage(message: message)
-        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
-        return isFromCurrentSender ? layout.outgoingMessagePadding : layout.incomingMessagePadding
-    }
-
-    internal func messageContainerMaxWidth(for message: MessageType, at indexPath: IndexPath) -> CGFloat {
-        let avatarWidth = avatarSize(for: message, at: indexPath).width
-        let messagePadding = messageContainerPadding(for: message, at: indexPath)
-
-        switch message.data {
-        case .text, .attributedText:
-            let messageInsets = messageLabelInsets(for: message, at: indexPath)
-            return itemWidth - avatarWidth - messagePadding.horizontal - messageInsets.horizontal
-        default:
-            return itemWidth - avatarWidth - messagePadding.horizontal
-        }
-    }
-
-    internal func messageContainerSize(for message: MessageType, at indexPath: IndexPath) -> CGSize {
-        let maxWidth = messageContainerMaxWidth(for: message, at: indexPath)
-
-        var messageContainerSize: CGSize = .zero
-
-        switch message.data {
-        case .text(let text):
-            let messageInsets = messageLabelInsets(for: message, at: indexPath)
-            let attributedText = NSAttributedString(string: text, attributes: [.font: messageLabelFont])
-            messageContainerSize = labelSize(for: attributedText, considering: maxWidth)
-            messageContainerSize.width += messageInsets.horizontal
-            messageContainerSize.height += messageInsets.vertical
-        case .attributedText(let text):
-            let messageInsets = messageLabelInsets(for: message, at: indexPath)
-            messageContainerSize = labelSize(for: text, considering: maxWidth)
-            messageContainerSize.width += messageInsets.horizontal
-            messageContainerSize.height += messageInsets.vertical
-        case .emoji(let text):
-            let messageInsets = messageLabelInsets(for: message, at: indexPath)
-            let attributedText = NSAttributedString(string: text, attributes: [.font: emojiLabelFont])
-            messageContainerSize = labelSize(for: attributedText, considering: maxWidth)
-            messageContainerSize.width += messageInsets.horizontal
-            messageContainerSize.height += messageInsets.vertical
-        case .photo, .video:
-            let width = messagesLayoutDelegate.widthForMedia(message: message, at: indexPath, with: maxWidth, in: messagesCollectionView)
-            let height = messagesLayoutDelegate.heightForMedia(message: message, at: indexPath, with: maxWidth, in: messagesCollectionView)
-            messageContainerSize = CGSize(width: width, height: height)
-        case .location:
-            let width = messagesLayoutDelegate.widthForLocation(message: message, at: indexPath, with: maxWidth, in: messagesCollectionView)
-            let height = messagesLayoutDelegate.heightForLocation(message: message, at: indexPath, with: maxWidth, in: messagesCollectionView)
-            messageContainerSize = CGSize(width: width, height: height)
-        case .custom:
-            fatalError(MessageKitError.customDataUnresolvedSize)
-        }
-
-        return messageContainerSize
-    }
-
-    // MARK: - Cell Size
-
-    internal func cellContentHeight(for message: MessageType, at indexPath: IndexPath) -> CGFloat {
-        let attributes = intermediateLayoutAttributes(for: message, at: indexPath)
-        return attributes.itemHeight
-    }
-
-    /// Returns the size for the `MessageCollectionViewCell` at a given `IndexPath`
-    /// considering all of the cell's content.
-    ///
-    /// - Parameters:
-    ///   - indexPath: The `IndexPath` of the cell.
-    open func sizeForItem(at indexPath: IndexPath) -> CGSize {
-        let message = messagesDataSource.messageForItem(at: indexPath, in: messagesCollectionView)
-        let itemHeight = cellContentHeight(for: message, at: indexPath)
-        return CGSize(width: itemWidth, height: itemHeight)
-    }
-}
-
-// MARK: - MessagesCollectionViewLayoutAttributes
-
-extension MessagesCollectionViewFlowLayout {
 
     private func intermediateLayoutAttributes(for message: MessageType, at indexPath: IndexPath) -> MessageIntermediateLayoutAttributes {
 
@@ -497,6 +218,287 @@ extension MessagesCollectionViewFlowLayout {
             break
         }
     }
+
+    // MARK: - Layout Invalidation
+
+    open override func shouldInvalidateLayout(forBoundsChange newBounds: CGRect) -> Bool {
+        if collectionView?.bounds.width != newBounds.width {
+            removeAllCachedAttributes()
+            return true
+        } else {
+            return false
+        }
+    }
+
+    open override func invalidationContext(forBoundsChange newBounds: CGRect) -> UICollectionViewLayoutInvalidationContext {
+        let context = super.invalidationContext(forBoundsChange: newBounds)
+        guard let flowLayoutContext = context as? UICollectionViewFlowLayoutInvalidationContext else { return context }
+        flowLayoutContext.invalidateFlowLayoutDelegateMetrics = shouldInvalidateLayout(forBoundsChange: newBounds)
+        return flowLayoutContext
+    }
+
+    /// Removes the cached layout information for all `MessageType`s.
+    public func removeAllCachedAttributes() {
+        textMessageAttributesCache.removeAllObjects()
+        attributedTextMessageAttributesCache.removeAllObjects()
+        emojiMessageAttributesCache.removeAllObjects()
+        photoMessageAttributesCache.removeAllObjects()
+        videoMessageAttributesCache.removeAllObjects()
+        locationMessageAttributesCache.removeAllObjects()
+        customMessageAttributesCache.removeAllObjects()
+    }
+
+    public func removeAllCachedAttributes(for messageType: MessageData) {
+        let cache = attributeCache(for: messageType)
+        cache.removeAllObjects()
+    }
+
+    public func removeCachedAttributes(for message: MessageType) {
+        let cache = attributeCache(for: message.data)
+        cache.removeObject(forKey: message.messageId as NSString)
+    }
+
+    public func removeCachedAttributes(for messageId: String) {
+        textMessageAttributesCache.removeObject(forKey: messageId as NSString)
+        attributedTextMessageAttributesCache.removeObject(forKey: messageId as NSString)
+        emojiMessageAttributesCache.removeObject(forKey: messageId as NSString)
+        photoMessageAttributesCache.removeObject(forKey: messageId as NSString)
+        videoMessageAttributesCache.removeObject(forKey: messageId as NSString)
+        locationMessageAttributesCache.removeObject(forKey: messageId as NSString)
+        customMessageAttributesCache.removeObject(forKey: messageId as NSString)
+    }
+
+    @objc
+    private func handleOrientationChange(_ notification: Notification) {
+        removeAllCachedAttributes()
+        invalidateLayout()
+    }
+
+    // MARK: - Avatar Size
+
+    internal func avatarPosition(for message: MessageType, at indexPath: IndexPath) -> AvatarPosition {
+        let layout = layoutForMessage(message: message)
+        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
+        var position = isFromCurrentSender ? layout.outgoingAvatarPosition : layout.incomingAvatarPosition
+
+        switch position.horizontal {
+        case .cellTrailing, .cellLeading:
+            break
+        case .natural:
+            position.horizontal = isFromCurrentSender ? .cellTrailing : .cellLeading
+        }
+        return position
+    }
+
+    internal func avatarSize(for message: MessageType, at indexPath: IndexPath) -> CGSize {
+        let layout = layoutForMessage(message: message)
+        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
+        return isFromCurrentSender ? layout.outgoingAvatarSize : layout.incomingAvatarSize
+    }
+
+    // MARK: - Cell Top Label Sizing
+
+    /// Returns the size of the `cellTopLabel` in a `MessageCollectionViewCell`
+    /// for a given `MessageType` and `IndexPath`.
+    /// - Parameters:
+    ///   - message: The `MessageType` object for the cell.
+    ///   - indexPath: The `IndexPath` for the cell.
+    open func cellTopLabelSize(for message: MessageType, at indexPath: IndexPath) -> CGSize {
+        let text = messagesDataSource.cellTopLabelAttributedText(for: message, at: indexPath)
+        guard let topLabelText = text else { return .zero }
+        let maxWidth = cellTopLabelMaxWidth(for: message, at: indexPath)
+        return labelSize(for: topLabelText, considering: maxWidth)
+    }
+
+    internal func cellTopLabelAlignment(for message: MessageType, at indexPath: IndexPath) -> LabelAlignment {
+        let layout = layoutForMessage(message: message)
+        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
+        return isFromCurrentSender ? layout.outgoingCellTopLabelAlignment : layout.incomingCellTopLabelAlignment
+    }
+
+    internal func cellTopLabelMaxWidth(for message: MessageType, at indexPath: IndexPath) -> CGFloat {
+        let alignment = cellTopLabelAlignment(for: message, at: indexPath)
+        let position = avatarPosition(for: message, at: indexPath)
+        let avatarWidth = avatarSize(for: message, at: indexPath).width
+        let containerSize = messageContainerSize(for: message, at: indexPath)
+        let messagePadding = messageContainerPadding(for: message, at: indexPath)
+
+        let avatarHorizontal = position.horizontal
+        let avatarVertical = position.vertical
+
+        switch (alignment, avatarHorizontal) {
+
+        case (.cellLeading, _), (.cellTrailing, _):
+            let width = itemWidth - alignment.insets.horizontal
+            return avatarVertical != .cellTop ? width : width - avatarWidth
+
+        case (.cellCenter, _):
+            let width = itemWidth - alignment.insets.horizontal
+            return avatarVertical != .cellTop ? width : width - (avatarWidth * 2)
+
+        case (.messageTrailing, .cellLeading):
+            let width = containerSize.width + messagePadding.left - alignment.insets.horizontal
+            return avatarVertical == .cellTop ? width : width + avatarWidth
+
+        case (.messageLeading, .cellTrailing):
+            let width = containerSize.width + messagePadding.right - alignment.insets.horizontal
+            return avatarVertical == .cellTop ? width : width + avatarWidth
+
+        case (.messageLeading, .cellLeading):
+            return itemWidth - avatarWidth - messagePadding.left - alignment.insets.horizontal
+
+        case (.messageTrailing, .cellTrailing):
+            return itemWidth - avatarWidth - messagePadding.right - alignment.insets.horizontal
+
+        case (_, .natural):
+            fatalError(MessageKitError.avatarPositionUnresolved)
+        }
+    }
+
+    // MARK: - Cell Bottom Label Sizing
+
+    /// Returns the size of the `cellBottomLabel` in a `MessageCollectionViewCell`
+    /// for a given `MessageType` and `IndexPath`.
+    /// - Parameters:
+    ///   - message: The `MessageType` object for the cell.
+    ///   - indexPath: The `IndexPath` for the cell.
+    open func cellBottomLabelSize(for message: MessageType, at indexPath: IndexPath) -> CGSize {
+        let text = messagesDataSource.cellBottomLabelAttributedText(for: message, at: indexPath)
+        guard let bottomLabelText = text else { return .zero }
+        let maxWidth = cellBottomLabelMaxWidth(for: message, at: indexPath)
+        return labelSize(for: bottomLabelText, considering: maxWidth)
+    }
+
+    internal func cellBottomLabelAlignment(for message: MessageType, at indexPath: IndexPath) -> LabelAlignment {
+        let layout = layoutForMessage(message: message)
+        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
+        return isFromCurrentSender ? layout.outgoingCellBottomLabelAlignment : layout.incomingCellBottomLabelAlignment
+    }
+
+    internal func cellBottomLabelMaxWidth(for message: MessageType, at indexPath: IndexPath) -> CGFloat {
+
+        let alignment = cellBottomLabelAlignment(for: message, at: indexPath)
+        let avatarWidth = avatarSize(for: message, at: indexPath).width
+        let containerSize = messageContainerSize(for: message, at: indexPath)
+        let messagePadding = messageContainerPadding(for: message, at: indexPath)
+        let position = avatarPosition(for: message, at: indexPath)
+
+        let avatarHorizontal = position.horizontal
+        let avatarVertical = position.vertical
+
+        switch (alignment, avatarHorizontal) {
+
+        case (.cellLeading, _), (.cellTrailing, _):
+            let width = itemWidth - alignment.insets.horizontal
+            return avatarVertical != .cellBottom ? width : width - avatarWidth
+
+        case (.cellCenter, _):
+            let width = itemWidth - alignment.insets.horizontal
+            return avatarVertical != .cellBottom ? width : width - (avatarWidth * 2)
+
+        case (.messageTrailing, .cellLeading):
+            let width = containerSize.width + messagePadding.left - alignment.insets.horizontal
+            return avatarVertical == .cellBottom ? width : width + avatarWidth
+
+        case (.messageLeading, .cellTrailing):
+            let width = containerSize.width + messagePadding.right - alignment.insets.horizontal
+            return avatarVertical == .cellBottom ? width : width + avatarWidth
+
+        case (.messageLeading, .cellLeading):
+            return itemWidth - avatarWidth - messagePadding.left - alignment.insets.horizontal
+
+        case (.messageTrailing, .cellTrailing):
+            return itemWidth - avatarWidth - messagePadding.right - alignment.insets.horizontal
+
+        case (_, .natural):
+            fatalError(MessageKitError.avatarPositionUnresolved)
+        }
+    }
+
+    // MARK: - Message Container Sizing
+
+    /// Returns the size of the `MessageContainerView` in a `MessageCollectionViewCell`
+    /// for a given `MessageType` and `IndexPath`.
+    /// - Parameters:
+    ///   - message: The `MessageType` object for the cell.
+    ///   - indexPath: The `IndexPath` for the cell.
+    open func messageContainerSize(for message: MessageType, at indexPath: IndexPath) -> CGSize {
+        let maxWidth = messageContainerMaxWidth(for: message, at: indexPath)
+
+        var messageContainerSize: CGSize = .zero
+
+        switch message.data {
+        case .text(let text):
+            let messageInsets = messageLabelInsets(for: message, at: indexPath)
+            let attributedText = NSAttributedString(string: text, attributes: [.font: messageLabelFont])
+            messageContainerSize = labelSize(for: attributedText, considering: maxWidth)
+            messageContainerSize.width += messageInsets.horizontal
+            messageContainerSize.height += messageInsets.vertical
+        case .attributedText(let text):
+            let messageInsets = messageLabelInsets(for: message, at: indexPath)
+            messageContainerSize = labelSize(for: text, considering: maxWidth)
+            messageContainerSize.width += messageInsets.horizontal
+            messageContainerSize.height += messageInsets.vertical
+        case .emoji(let text):
+            let messageInsets = messageLabelInsets(for: message, at: indexPath)
+            let attributedText = NSAttributedString(string: text, attributes: [.font: emojiLabelFont])
+            messageContainerSize = labelSize(for: attributedText, considering: maxWidth)
+            messageContainerSize.width += messageInsets.horizontal
+            messageContainerSize.height += messageInsets.vertical
+        case .photo, .video:
+            let width = messagesLayoutDelegate.widthForMedia(message: message, at: indexPath, with: maxWidth, in: messagesCollectionView)
+            let height = messagesLayoutDelegate.heightForMedia(message: message, at: indexPath, with: maxWidth, in: messagesCollectionView)
+            messageContainerSize = CGSize(width: width, height: height)
+        case .location:
+            let width = messagesLayoutDelegate.widthForLocation(message: message, at: indexPath, with: maxWidth, in: messagesCollectionView)
+            let height = messagesLayoutDelegate.heightForLocation(message: message, at: indexPath, with: maxWidth, in: messagesCollectionView)
+            messageContainerSize = CGSize(width: width, height: height)
+        case .custom:
+            fatalError(MessageKitError.customDataUnresolvedSize)
+        }
+
+        return messageContainerSize
+    }
+
+    internal func messageLabelInsets(for message: MessageType, at indexPath: IndexPath) -> UIEdgeInsets {
+        guard let layout = layoutForMessage(message: message) as? TextMessageLayout else {
+            fatalError("Layout is not TextMessageLayout")
+        }
+        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
+        return isFromCurrentSender ? layout.outgoingMessageLabelInsets : layout.incomingMessageLabelInsets
+    }
+
+    internal func messageContainerPadding(for message: MessageType, at indexPath: IndexPath) -> UIEdgeInsets {
+        let layout = layoutForMessage(message: message)
+        let isFromCurrentSender = messagesDataSource.isFromCurrentSender(message: message)
+        return isFromCurrentSender ? layout.outgoingMessagePadding : layout.incomingMessagePadding
+    }
+
+    internal func messageContainerMaxWidth(for message: MessageType, at indexPath: IndexPath) -> CGFloat {
+        let avatarWidth = avatarSize(for: message, at: indexPath).width
+        let messagePadding = messageContainerPadding(for: message, at: indexPath)
+
+        switch message.data {
+        case .text, .attributedText:
+            let messageInsets = messageLabelInsets(for: message, at: indexPath)
+            return itemWidth - avatarWidth - messagePadding.horizontal - messageInsets.horizontal
+        default:
+            return itemWidth - avatarWidth - messagePadding.horizontal
+        }
+    }
+
+    // MARK: - Cell Sizing
+
+    /// Returns the size of a `MessageCollectionViewCell` at a given `IndexPath`
+    /// considering all of the cell's content.
+    ///
+    /// - Parameters:
+    ///   - indexPath: The `IndexPath` for the cell.
+    open func sizeForItem(at indexPath: IndexPath) -> CGSize {
+        let message = messagesDataSource.messageForItem(at: indexPath, in: messagesCollectionView)
+        let attributes = intermediateLayoutAttributes(for: message, at: indexPath)
+        return CGSize(width: itemWidth, height: attributes.itemHeight)
+    }
 }
 
 // MARK: - Helpers
@@ -541,7 +543,6 @@ extension MessagesCollectionViewFlowLayout {
         }
     }
 
-    /// Convenience property for accessing the layout object's `MessagesCollectionView`.
     internal var messagesCollectionView: MessagesCollectionView {
         guard let messagesCollectionView = collectionView as? MessagesCollectionView else {
             fatalError(MessageKitError.layoutUsedOnForeignType)
@@ -549,7 +550,6 @@ extension MessagesCollectionViewFlowLayout {
         return messagesCollectionView
     }
 
-    /// Convenience property for unwrapping the `MessagesCollectionView`'s `MessagesDataSource`.
     internal var messagesDataSource: MessagesDataSource {
         guard let messagesDataSource = messagesCollectionView.messagesDataSource else {
             fatalError(MessageKitError.nilMessagesDataSource)
@@ -557,60 +557,10 @@ extension MessagesCollectionViewFlowLayout {
         return messagesDataSource
     }
 
-    /// Convenience property for unwrapping the `MessagesCollectionView`'s `MessagesLayoutDelegate`.
     internal var messagesLayoutDelegate: MessagesLayoutDelegate {
         guard let messagesLayoutDelegate = messagesCollectionView.messagesLayoutDelegate else {
             fatalError(MessageKitError.nilMessagesLayoutDelegate)
         }
         return messagesLayoutDelegate
     }
-}
-
-class MessageIntermediateLayoutAttributes {
-
-    var avatarSize: CGSize = .zero
-    var avatarPosition = AvatarPosition(vertical: .cellBottom)
-
-    var messageContainerSize: CGSize = .zero
-    var messageContainerPadding: UIEdgeInsets = .zero
-    var messageLabelFont: UIFont = UIFont.preferredFont(forTextStyle: .body)
-    var messageLabelInsets: UIEdgeInsets = .zero
-
-    var topLabelAlignment = LabelAlignment.cellCenter(.zero)
-    var topLabelSize: CGSize = .zero
-
-    var bottomLabelAlignment = LabelAlignment.cellCenter(.zero)
-    var bottomLabelSize: CGSize = .zero
-
-    lazy var itemHeight: CGFloat = {
-        let avatarVerticalPosition = avatarPosition.vertical
-        let avatarHeight = avatarSize.height
-        let messageContainerHeight = messageContainerSize.height
-        let bottomLabelHeight = bottomLabelSize.height
-        let topLabelHeight = topLabelSize.height
-        let messageVerticalPadding = messageContainerPadding.vertical
-
-        var cellHeight: CGFloat = 0
-
-        switch avatarVerticalPosition {
-        case .cellTop:
-            cellHeight += max(avatarHeight, topLabelHeight)
-            cellHeight += bottomLabelHeight
-            cellHeight += messageContainerHeight
-            cellHeight += messageVerticalPadding
-        case .cellBottom:
-            cellHeight += max(avatarHeight, bottomLabelHeight)
-            cellHeight += topLabelHeight
-            cellHeight += messageContainerHeight
-            cellHeight += messageVerticalPadding
-        case .messageTop, .messageCenter, .messageBottom:
-            cellHeight += max(avatarHeight, messageContainerHeight)
-            cellHeight += messageVerticalPadding
-            cellHeight += topLabelHeight
-            cellHeight += bottomLabelHeight
-        }
-
-        return cellHeight
-    }()
-
 }
